refactored sortable lists
