+++ server/db.json
@@ -15,17 +15,58 @@
       "friendRequests": [],
       "travels": [
         {
+          "id": 1696690300138,
+          "name": "tri123",
+          "dateStart": "2023.10.07",
+          "dateEnd": "2023.10.07",
+          "members": [],
+          "backlog": [],
+          "groups": []
+        },
+        {
+          "id": 1696690304858,
+          "name": "fasd123",
+          "dateStart": "2023.10.07",
+          "dateEnd": "2023.10.07",
+          "members": [],
+          "backlog": [],
+          "groups": []
+        },
+        {
           "id": 1696687791519,
           "name": "new trip",
           "dateStart": "2023.10.07",
           "dateEnd": "2023.10.07",
           "members": [],
-          "backlog": [],
+          "backlog": [
+            {
+              "id": 1696689795901,
+              "name": "678"
+            },
+            {
+              "id": 1696689789085,
+              "name": "123"
+            },
+            {
+              "id": 1696689792596,
+              "name": "345"
+            }
+          ],
           "groups": [
+            {
+              "id": 1696690045658,
+              "title": "new group12",
+              "cards": []
+            },
             {
               "id": 1696687802657,
               "title": "123",
               "cards": [
+                {
+                  "id": 1696690808204,
+                  "title": "123",
+                  "description": "adsf"
+                },
                 {
                   "id": 1696687812670,
                   "title": "qwe",
