+++ src/components/CardsList/CardItem/CardItem.tsx
@@ -2,8 +2,7 @@ import { FC, memo } from 'react';
 import { useAppDispatch } from 'hooks/redux';
 import { userSlice } from 'redux/reducers/userSlice';
 import { ICard } from 'types/types';
-import { useSortable } from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
+import { SortableItem } from 'components/SortableItem/SortableItem';
 import classnames from 'classnames';
 import classes from './CardItem.module.css';
 
@@ -14,7 +13,6 @@ interface CardItemProps {
 	className?: string;
 }
 
-
 export const CardItem:FC<CardItemProps> = memo((props) => {
 	const {
 		card,
@@ -23,18 +21,6 @@ export const CardItem:FC<CardItemProps> = memo((props) => {
 		className,
 	} = props;
 	const dispatch = useAppDispatch();
-	const { 
-		attributes,
-    	listeners,
-    	setNodeRef,
-    	transform,
-    	transition
-   	} = useSortable({id: card.id});
-
-   	const style = {
- 		transform: CSS.Translate.toString(transform),
- 		transition
-	};
 
 	const handleDeleteClick = ():void => {
 		dispatch(userSlice.actions.deleteCard({
@@ -45,22 +31,22 @@ export const CardItem:FC<CardItemProps> = memo((props) => {
 	};
 
 	return (
-		<li 
-			className={classnames(classes.CardItem, className)}
-			style={style}
-			ref={setNodeRef}
-			{...attributes} 
-			{...listeners}
-		>
-			<div className={classes.card__left}>
-				<h3 className={classes.card__title}>{card.title}</h3>
-			</div>
-			<span className={classes.card__separator}></span>
-			<div className={classes.card__text}>{card.description}</div>
-			<button 
-				onClick={handleDeleteClick}
-				className={classes.delete}
-			>&#10005;</button>
-		</li>
+		<SortableItem id={card.id}>
+			<li className={classnames(classes.CardItem, className)}>
+				<div className={classes.card__left}>
+					<h3 className={classes.card__title}>{card.title}</h3>
+				</div>
+				<span className={classes.card__separator}></span>
+				<div className={classes.card__text}>
+					{card.description}
+				</div>
+				<button 
+					onClick={handleDeleteClick}
+					className={classes.delete}
+				>
+					&#10005;
+				</button>
+			</li>
+		</SortableItem>
 	);
 });
