+++ src/components/Groups/GroupItem/GroupItem.tsx
@@ -5,6 +5,7 @@ import { useAppDispatch } from 'hooks/redux';
 import { userSlice } from 'redux/reducers/userSlice';
 import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
+
 import CardCreateForm from 'components/CreateForms/CardCreateForm/CardCreateForm';
 import CardsList from 'components/CardsList/CardsList';
 import Popup from 'components/Popup/Popup';
