+++ src/components/SortableItem/SortableItem.tsx
@@ -1,0 +1,34 @@
+import React, { FC, ReactNode } from "react";
+import { useSortable } from "@dnd-kit/sortable";
+import { CSS } from "@dnd-kit/utilities";
+
+interface SortableItemProps {
+    id: number;
+    children: ReactNode;
+}
+
+export const SortableItem: FC<SortableItemProps> = ({id, children}) => {
+    const {
+        attributes,
+        listeners,
+        setNodeRef,
+        transform,
+        transition
+    } = useSortable({ id: id });
+
+    const style = {
+        transform: CSS.Transform.toString(transform),
+        transition,
+    };
+
+    return (
+        <div
+            ref={setNodeRef}
+            style={style}
+            {...attributes}
+            {...listeners}
+        >
+            {children}
+        </div>
+    );
+}
