+++ src/components/travel/TravelItem/TravelItem.module.css
@@ -1,0 +1,130 @@
+.list__item {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+
+
+	padding: 15px 10px;
+	margin: 10px 0;
+	border-radius: 8px;
+	box-shadow: -1px 8px 25px -12px rgba(0,0,0,0.75);
+}
+
+.list__item:hover {
+	cursor: grab;
+}
+
+.list__item_details {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+
+
+	padding: 15px 0;
+	margin: 10px 0;
+	border-radius: 8px;
+}
+
+.item__title {
+	width: 48%;
+	margin-right: 10px;
+}
+
+.item__link {
+	text-decoration: none;
+	color: inherit;
+	font-size: 14px;
+}
+
+.item__link_details {
+	text-decoration: none;
+	color: inherit;
+	font-size: 22px;
+	font-weight: 500;
+
+	pointer-events: none;
+}
+
+.item__link:hover {
+	text-decoration: underline;
+}
+
+
+.item__input {
+	width: 44%;
+	padding: 10px;
+	font-size: 14px;
+	margin-right: 10px;
+}
+
+.item__date {
+	padding: 0px;
+	width: 47%;
+
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+}
+
+.date__separator {
+	font-size: 20px;
+	font-weight: 500;
+
+	margin: 0 6px;
+	height: 100%;
+}
+
+.item__actions {
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content: end;
+
+	width: fit-content;
+
+	gap: 5px;
+	padding: 0 3px;
+}
+
+.edit,
+.delete {
+	width: 32px;
+	height: 37px;
+	padding: 2px 0;
+	font-size: 0;
+
+	border: none;
+	background-position: center;
+	background-repeat: no-repeat;
+	background-size: contain;
+	color: transparent;
+}
+
+.edit {
+	background-color: var(--color-lightblue);
+
+	background-image: url('assets/icons/edit_white.svg');
+}
+
+.delete {
+	background-color: var(--color-red);
+
+	background-image: url('assets/icons/delete.svg');
+}
+
+.save {
+	width: 32px;
+	height: 37px;
+	padding: 2px 0;
+
+	font-weight: 500;
+	font-size: 16px;
+
+	border: none;
+}
+
+.edit:hover,
+.delete:hover,
+.save:hover {
+	cursor: pointer;
+}
