+++ src/pages/TravelDetailsPage/TravelDetailsPage.tsx
@@ -1,12 +1,12 @@
-import {FC} from 'react';
-import {useParams} from 'react-router-dom';
-import {useAppSelector} from 'hooks/redux';
+import { FC } from 'react';
+import { useParams } from 'react-router-dom';
+import { useAppSelector } from 'hooks/redux';
 import { selectTravelById } from 'redux/selectors/selectors';
+import { TravelItem } from 'components/travel/TravelItem/TravelItem';
 import NavBar from 'components/Navbar/NavBar';
 import Members from 'components/Members/Members';
 import BacklogList from 'components/BacklogList/BacklogList';
 import Groups from 'components/Groups/Groups';
-import TravelItem from 'components/travel/TravelItem';
 import classes from './travel-details.module.css';
 
 const TravelDetailsPage: FC = () => {
